Replace input validation alert to a modal. Move the UI components to UI folder
